Big Update :) Added firebase auth & streamlined sign in process.
Users can now create an account or sign in with an existing account through the firebase Auth SDK. Implemented return buttons for navigation views to make moving through them easier. Added a "Sign Out" button to the Account page which takes users out of the app and back to the welcome screen.