UNSTABLE / V UGLY Implementation of creating an event
Lots of new changes. Updated the button UI on MapView to look a bit better. Made the Sign Out button in profileView dynamically colored to show up in dark/light mode.

Also started work on the create event feature, and now the app can actually create an event struct, AND upload it to the firestore database!!!! It's super ugly code so I'm worried abt long term, but for now it works.

Still need to make the forwardGeocoding work.